Convert everything to use FCBs.
The Standard Benchmark is now up to 7m54s, from 7m10s. The extra time is mostly
spent extending files as because we can't use write-only streams because of
needing to read from the file for the buffering, we don't get the automatic
disk allocation which ADFS does for us.